Implement "prev" navigation
This commit implements navigation to previous page of results via an
`action` parameter to distinguish between the forward (next) and
backwared (previous) navigation cases.

Assuming we have an auto-incrementing identifier and we are sorting in
descending order, suppose we have the following data:

```
[10, 9, 8, 7, 6, 5, 4, 3, 2, 1]
```

...and a `pageSize` of, say 4.

The first page should contain [10, 9], with the following links:

- `next`: /items?pageSize=4&lastSeen=9
- `self`: /items?pageSize=4

Since this is the first page, we don't expect a `prev` link.

The second page, should contain [8,7], with:

- `next`: /items?pageSize=4&lastSeen=7
- `prev`: /items?pageSize=4&lastSeen=9&action=previous
- `self`: /items?pageSize=4&lastSeen=9

Peforming a `GET` on the `prev` link, should get back to the first page
(same data, same links).